MyAppTemplate/requirements.txt
streamlit>=0.56.0
numpy
pandas
django-environ
pysolr
altair
matplotlib
plotly
plotly_express

